[core] Move as_hal methods to their own file
